\name{boxPlotX}

\alias{boxPlotX}

\title{
Box-Whisker plot.
}

\encoding{latin1}

\description{
This function generates plot by iClick.VisOneReturns. 
}
\usage{
boxPlotX(X, col = "indianred2", title = TRUE)
}
\arguments{
  \item{X}{
A timeSeries object, single time series returns.
}
  \item{col}{
String for color.
}
  \item{title}{
Whether to generate title of graph.
}

}

\details{
This function is an internal function of iClick GUI, which is executed on iClick.VisOneReturns GUI.
}

\value{
Plot a graph
}

\author{
Ho Tsung-wu <tsungwu@mail.shu.edu.tw>
}

\seealso{
Functions in fBasics.
}

