\name{data-sets}

\alias{dataSets}

\alias{assetReturns}

\alias{IBM}

\alias{FFplusMOM}

\alias{world20}

\alias{returnsDaily24}

\title{Data Sets}

\encoding{latin1}

\description{

assetReturns contains DJ component stocks returns data. world20 contains index returns of 20 countries.

}


\usage{
data(assetReturns)
data(world20)
}


\value{

  assetReturns is a time series object of package \code{xts}.
	world20 is index returns time series data.
}


\keyword{datasets}
