\name{drawups}
\alias{drawups}

\title{
Calculate Drawup Returns for Drawup Plot
}

\encoding{latin1}

\description{
This function calculates drawup returns for plotting.
}
\usage{
drawups(x)
}

\arguments{
  \item{x}{
A timeSeries object, single time series returns.
}
}
\details{
This function is an internal function for drawplot of iClick GUI, which is executed on iClick.VisOneReturns GUI.
}
\value{
Returns of draw up periods.
}

\author{
Ho Tsung-wu <tsungwu@mail.shu.edu.tw>
}


