\name{IBM}
\alias{IBM}
\docType{data}
\title{
Daily Price Data of IBM
}

\encoding{latin1}

\description{
Daily price data of IBM, 2007/1/3~2015/10/29
}
\usage{data("IBM")}
\format{
  A data frame with 2223 observations on the following 6 variables.
  \describe{
    \item{\code{Date}}{a date string}
    \item{\code{IBM.OPEN}}{a numeric vector}
    \item{\code{IBM.HIGH}}{a numeric vector}
    \item{\code{IBM.LOW}}{a numeric vector}
    \item{\code{IBM.CLOSE}}{a numeric vector}
    \item{\code{IBM.VOLUME}}{a numeric vector}
    \item{\code{IBM.ADJ.CLOSE}}{a numeric vector}    
  }
}


\details{
Daily stock price data of IBM.
}
\source{
Yahoo finance.
}

