\name{iClick-package}

\alias{iClick-package}

\alias{iClick}

\docType{package}
\encoding{latin1}
\title{
A Button-based GUI for Financial and Economic Data Analysis
}


\description{A Output GUI designed to simplfy the use of R packages and functions by clicking.
}


\author{Ho Tsung-wu <tsungwu@mail.shu.edu.tw>

Maintainer: Ho Tsung-wu  <tsungwu@mail.shu.edu.tw>
}

