\name{returnsDaily24}
\alias{returnsDaily24}
\docType{data}
\encoding{latin1}
\title{
Daily Returns Data of 24 Markets
}


\description{
Daily returns data of 24 world national market index, 2001/1/03~2013/9/24
}
\usage{data("returnsDaily24")}
\format{
  A data frame with 3320 observations on the following 24 variables.
  \describe{
    \item{\code{Dates}}{Time string}
    \item{\code{AEX}}{a numeric vector of national market}
    \item{\code{AORD}}{a numeric vector of national market}
    \item{\code{ATX}}{a numeric vector of national market}
    \item{\code{BFX}}{a numeric vector of national market}
    \item{\code{BVSP}}{a numeric vector of national market}
    \item{\code{FCHI}}{a numeric vector of national market}
    \item{\code{FTSE}}{a numeric vector of national market}
    \item{\code{FTSEMIB.MI}}{a numeric vector of national market}
    \item{\code{GD.AT}}{a numeric vector of national market}
    \item{\code{GDAXI}}{a numeric vector of national market}
    \item{\code{GSPC}}{a numeric vector of national market}
    \item{\code{GSPTSE}}{a numeric vector of national market}
    \item{\code{HSI}}{a numeric vector of national market}
    \item{\code{JKSE}}{a numeric vector of national market}
    \item{\code{KLSE}}{a numeric vector of national market}
    \item{\code{KS11}}{a numeric vector of national market}
    \item{\code{MERV}}{a numeric vector of national market}
    \item{\code{MXX}}{a numeric vector of national market}
    \item{\code{N225}}{a numeric vector of national market}
    \item{\code{OMX}}{a numeric vector of national market}
    \item{\code{SSEC}}{a numeric vector of national market}
    \item{\code{SSMI}}{a numeric vector of national market}
    \item{\code{STI}}{a numeric vector of national market}
    \item{\code{TWII}}{a numeric vector of national market}
    
  }
}
\details{
Daily stock returns of 24 world national markets.
}
\source{
Yahoo finance.
}


