\name{VIF_no}
\alias{VIF_no}

\title{
VIF test for mullticolinearity
}

\encoding{latin1}

\description{
This function tests for mullticolinearity.
}
\usage{
VIF_no(obj)
}
\arguments{
  \item{obj}{
A lm object
}


}

\details{
This function is an internal function of iClick GUI, which is executed on iClick.lm GUI.
}

\value{
Test statistic
}

\author{
Ho Tsung-wu <tsungwu@ntnu.edu.tw>, College of Management, National Taiwan Normal University
}

