\name{IBM}
\alias{IBM}
\docType{data}
\title{
Daily Price Data of IBM
}

\encoding{latin1}

\description{
Daily price data of IBM, 2007/1/3~2015/10/29
}
\usage{data("IBM")}
\format{
  A data frame with 2223 observations on the following 6 variables.
  \describe{
    \item{\code{X}}{A date string}
    \item{\code{IBM.OPEN}}{A numeric vector}
    \item{\code{IBM.HIGH}}{A numeric vector}
    \item{\code{IBM.LOW}}{A numeric vector}
    \item{\code{IBM.CLOSE}}{A numeric vector}
    \item{\code{IBM.VOLUME}}{A numeric vector}
    \item{\code{IBM.ADJ.CLOSE}}{A numeric vector}
  }
}


\details{
Daily stock price data of IBM.
}
\source{
Yahoo finance.
}

